"init" doc: mention the configurability more clearly
Update the description for the '--initial-branch' option to mention
that the historical default has been `master` and it can be
configured to use any name.  When actual transition happens to
another fixed name, say 'main', the description would further be
updated to mention like:

	... fall back to the default name 'main', but it can be
	customized with the `init.defaultBranch` configuration
	variable.  Note that the default name historically was
	'master'.

in order to help people who wonder why more mature projects or
documentation tend not to say 'main'.

While we are at it, stop referring to "HEAD file", which was a
remnant from 2005 that talks too much about implementation detail.
The most important fact is that the current branch is set to an
unborn branch and not detached HEAD.

Signed-off-by: Junio C Hamano <gitster@pobox.com>